🐛 버그 수정중

+++ src/main/java/com/spring/mood/projectmvc/dto/requestDto/ShopItemRequestDto.java
@@ -3,6 +3,7 @@ package com.spring.mood.projectmvc.dto.requestDto;
 import lombok.*;
 import com.spring.mood.projectmvc.entity.ShopItem;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter @ToString
@@ -14,7 +15,7 @@ import java.time.LocalDateTime;
 public class ShopItemRequestDto {
     private String shopItemName;
     private String shopItemDesc;
-    private Double shopItemPrice;
+    private BigDecimal shopItemPrice;
     private String shopItemImg;
     private LocalDateTime shopItemDate;
     private long shopItemStock;
